Add multiprocessing example
Multiprocessing support is provided by the command line interface.
This is intended for calculating multiple spectra with different
parameters.

A user could also split the mesh and calculate the spectrum in each
submesh using this method, latter merging the returned Spectrum
instances. I do not plan to implement functionality for this, since I
don't think such an optimization is needed for a single spectrum
calculation.